Added old and new tests for `copy_and_compile.sh`.
I wrote new tests using the [`bats` tool](https://github.com/sstephenson/bats) with the intention of eventually replacing the old `shunit_2` tests. I also added the tarball directly to the project so that the lab doesn't depend on important files being in particular locations on a particular file system.