Removed the old `Pre-lab.md` file.
Since I've moved the pre-lab to its own repository, we don't need this
Markdown file here anymore.

In the future pre-labs I definitely think splitting them is important since
we'll want them to fork the pre-lab repository. It's less clear that it's
crucial here since they're (currently) not forking or "turning in" anything
for this pre-lab. I did it anyway just to be consistent.