Updated the write-up and tests for `extract_and_compile.sh`.
I renamed target script from `copy_and_compile.sh` to `extract_and_compile.sh` because we're not copying anymore.

I modified the tests and the write-up so that we now provide the target directory as the second argument and we assume it's already been constructed, instead of having our script create and delete it.

I also removed the bit where it wrote the results to a file, and instead just looked at the results as they come out on the command line (i.e., the `$output` variable in `bats`).